grabpatch - Patch to fix Grab problems
This patch fixes a potential problem with the Grab command. On some
systems, you may find the Grab command going 'out-to-lunch', or maybe it
returns, but doesn't appear to have done anything. This patch should
probably be applied on all systems, even if they don't exhibit the
problem, as it may be WM-dependent.