xv-3.10a.JPEG-patch - Patch to use version 6 of the JPEG library
John said:
You need this trivial patch if you intend to link xv with version 6 of
the Independent JPEG Group's JPEG library. This is a newer version of
the library than is included in the xv 3.10a distributions listed above.
You need the new library to read Progressive JPEG files.