xv-3.10a.TIFF-patch - Patch to use version 3.4 of the TIFF library
This patch is needed to link xv against newer versions of libtiff

John said:
You need this patch if you intend to link xv with version 3.4 of Sam
Leffler's TIFF library. This is a newer version of the library than is
included in the xv 3.10a distribution, and probably does something good.
(I haven't tried it out.)