mp-tiff-patch - Patch to read & display multi-page TIFF files
A cool one: like it says, this patch will let xv read and display
multi-page TIFF files. Multi-page TIFF files work just like multi-page
PostScript files: you use <PageUp> and <PageDown> to page through the
file, and p brings up a "go to page #" dialog box. Note that this is a
patch to the standard xv 3.10a distribution - it may or may not work in
combination with the libtiff 3.4 patch.