croppad.patch - Fix a crop problem
John said:
As pointed out by long-time Friend-of-XV Rick Dyson, the XV 3.10a does
an unexpected thing if you crop an image, and then 'pad' it. 3.10a would
effectively 'UnCrop' the image first, then pad it. There is a workaround
(save and reload the cropped image), but that's every bit as tedious as
it sounds.

This should do the trick, with any luck.