gifpatch - Patch to GIF-reading code
This patch fixes a bug in XVs gif-reading code.  The bug could conceivably
hit when viewing a number of interlaced GIFs.

John said:
While I can't say that I've ever seen the results of this bug
(obviously, or I would have fixed it), it's a bug nonetheless.  The bug
could show up when viewing multiple interlaced GIF files in one run of
XV, as the de-interlacer is not properly reinitialized on subsequent
runs.  Thanks to Mark Hanson for finding it.