exceeed_grab.patch - Patch to fix non-functional grab on some X servers
This patch fixes an inability to GRAB or display XWD files on the eXceed
X server running in 24-bits.  It may also fix GRAB problems on other X
servers.

The bits_per_item of the returned XImage is less than bits_per_pixel, which
is wrong, as far as I can tell.  (Hard to really know, as the X11 docs aren't
at *all* clear on the subject.)  I do know that it breaks my code, and the
netpbm code, and god only knows what else.

Maybe they're wrong, maybe I'm wrong - who knows.  In any event, this patch
fixes it, as far as XV is concerned.

John said:
xv would rather rudely fail to Grab (or display certain XWD files) on systems
running the eXceed X server in 24-bit mode.  The Grabs would return a
correctly-sized, but completely blank image.  This patch may well fix Grab
problems on other systems (and if so, I'd like to hear about it), and should
not break anything.   As such, it's recommended for all users.  Thanks to Nadim
Saeed for his efforts in squishing this bug.