windowid.patch - Patch to use XV as a Netscape plugin
I have added a "-windowid windowid" flag to xv so that you can choose an X
display to display xv window's on to. The advantage of this is
that you can use it now in conjunction with "plugger 3.2" which
lets you pick applications to use as plugins for Netscape.

In addition, I added a new "-rmode 10" - root mode 10, which just puts
a single copy of the picture in the upper left hand corner of
the DISPLAY. I had to do this so that it would look good in a browser.

The plugger home page is:
http://www.hubbe.net/~hubbe/plugger.html

The test page is:
http://www.hubbe.net/~hubbe/plugger/test.html

I tested using xv to DISPLAY a TIFF image, and a Sun raster image
into a frame of a Netscape browser.

Here is some of the plugger configuration code I used:

image/tiff: tiff,tif: TIFF image
image/x-tiff: tiff,tif: TIFF image
image/sun-raster: rs: SUN raster image
image/x-sun-raster: rs: SUN raster image
        : xv -rmode 10 -windowid $window $file -quit
image/png: png: PNG image
image/x-png: png: PNG image
        : xv -rmode 10 -windowid $window $file -quit

I have contacted the author the "Plugger 3.2" to let him know that
this new option will soon be available as a patch for xv, but I'd
like to put up the changes directly on your site.

I have included the patches with this email.  Please let me know

* if you can apply them with no problems

* if the code looks clean (up to your standards)

Note that if you just change the setup code slightly for plugger:
	: xv -windowid $window $file

You can have a fully functioning xv program in a Netscape browser.
You can even right click to bring up xv's controls, etc..

John said:
Contributed by David Meleedy, this patch adds a '-windowid #' command-line
option to xv, which was previously sorely lacking in command-line options. In
any event, the practical upshot of all this is that you can use xv in
conjunction with Plugger 3.2 to turn it into a Netscape plug-in.

Why would you want to do that, other than to win some sort of bet? Well, the big
deal is that by doing this, you can now get Netscape to display non-standard
image formats (TIFF, Rasterfile, PNG, etc.) right in a Netscape window, which is
certainly cooler than downloading them and firing up xv manually.  You can even
have a fully-functioning xv, schnauzer and all, running right inside a Netscape
window.   (Though I dare anyone to come up with a legitimate sporting purpose
for that!)